Separate wasm and native test functions
